✅Added Manage Listings Page & User Authorization

+++ app/Http/Controllers/ListingController.php
@@ -40,9 +40,12 @@ class ListingController extends Controller
             'description' => 'required'
         ]);
 
+        // Check if logo is uploaded & store it
         if($request->hasFile('logo')) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
+
+        $formFields['user_id'] = auth()->id();
         
         //Static method
         Listing::create($formFields);
@@ -61,6 +64,12 @@ class ListingController extends Controller
     // Update listing data
     public function update(Request $request, Listing $listing) {
         // dd($request->file('logo'));
+
+        //Make sure Logged in user is the owner of the listing
+        if($listing->user_id !==auth()->id()) {
+            abort(403, 'You are not authorized to edit this listing!');
+        }
+        
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -71,9 +80,13 @@ class ListingController extends Controller
             'description' => 'required'
         ]);
 
+        // Check if logo is uploaded & store it
         if($request->hasFile('logo')) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
+
+        //Adding a user id to the form fields
+        // $formFields['user_id'] = auth()->id;
         
         // Regular method
         $listing->update($formFields);
@@ -83,8 +96,18 @@ class ListingController extends Controller
 
     // Delete listing
     public function destroy(Listing $listing) {
+        //Make sure Logged in user is the owner of the listing
+        if($listing->user_id !==auth()->id()) {
+            abort(403, 'You are not authorized to edit this listing!');
+        }
+        
         $listing->delete();
 
         return redirect('/')->with('message', 'Listing deleted successfully!');
+    }
+
+     // Manage Listings
+     public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
